r2578: The attached patch adds a new module parameter fault_count which is updated whenever a fault is inserted - this allows test code using fault insertion to know for sure that faults have been simulated. I also fixed the formatting of the code to conform to the normal standard and also fixed a couple of warnings in trace statement that trace sector numbers using 64-bit formats but without the explicit cast to 64-bit size.
Also included is an update to the unit test script to use the new
variable and to remove a test that guaranteed creation of split-brain!

+++ drbd/drbd_req.c
@@ -84,7 +84,9 @@ void _print_rq_state(drbd_request_t *req, const char *txt)
 	     s & RQ_NET_SENT ? 's' : '-',
 	     s & RQ_NET_DONE ? 'd' : '-',
 	     s & RQ_NET_OK ? 'o' : '-',
-	     req->epoch, req->sector, req->size,
+	     req->epoch, 
+	     (unsigned long long)req->sector, 
+	     req->size,
 	     conns_to_name(mdev->state.conn));
 }
 
@@ -655,7 +657,9 @@ void _req_mod(drbd_request_t *req, drbd_req_event_t what)
 			     s & RQ_NET_SENT ? 's' : '-',
 			     s & RQ_NET_DONE ? 'd' : '-',
 			     s & RQ_NET_OK ? 'o' : '-',
-			     req->epoch, req->sector, req->size,
+			     req->epoch, 
+			     (unsigned long long)req->sector, 
+			     req->size,
 			     conns_to_name(mdev->state.conn));
 		}
 		D_ASSERT(req->rq_state & RQ_NET_SENT);
